% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave.R
\name{c_js}
\alias{c_js}
\title{c_js}
\usage{
c_js(x, y, concavity = 2, length_threshold = 1)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{concavity}{a relative measure of concavity. 1 results in a relatively detailed shape, Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{length_threshold}{when a segment length is under this threshold, it stops being considered for further detalization. Higher values result in simpler shapes.}
}
\value{
concave hull
}
\description{
'javascipt' implementation of concaveman, mostly from code of @joelgombin
}
