% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concavetest-package.R
\docType{package}
\name{concavetest-package}
\alias{concavetest}
\alias{concavetest-package}
\title{concavetest: Test 'libv8' and 'cpp' versions of concaveman algorithm}
\description{
Test 'libv8' and 'cpp' versions of concaveman algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/concavetest}
  \item Report bugs at \url{https://github.com/mpadge/concavetest/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Other contributors:
\itemize{
  \item Stanislaw Adaszewski (author of include concaveman-cpp code) [copyright holder]
}

}
\keyword{internal}
